requirements.txt
asgiref==3.5.2
colorama==0.4.5
distlib==0.3.6
Django==4.1
django-bootstrap4==22.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
Jinja2==3.1.2
pycparser==2.21
Pygments==2.13.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
